Beam/TextAnalysis/requirements.txt
google-cloud-firestore==1.9.0
nltk
gensim

